rolevate-interview/v2/requirements.txt
# LiveKit and AI Services
livekit-agents
livekit-plugins-openai
livekit-plugins-soniox
livekit-plugins-silero
livekit
livekit-plugins-elevenlabs

# HTTP and Async
aiohttp==3.9.1
python-dotenv==1.0.0

# AWS
boto3==1.34.0

# Configuration and Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Retry and Resilience
tenacity==8.2.3

# Logging
structlog==24.1.0

# Testing (dev)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
aioresponses==0.7.6

# Code Quality (dev)
black==23.12.0
ruff==0.1.9
mypy==1.7.1